[code quality] Add test file for base client
